Move AudioSource to its own file
