+++ lib/flutter_soloud.dart
@@ -2,6 +2,7 @@
 ///
 library flutter_soloud;
 
+export 'src/audio_source.dart';
 export 'src/enums.dart' hide PlayerErrors;
 export 'src/exceptions/exceptions.dart';
 export 'src/filter_params.dart';
